auth changes

+++ src/app/mixer/pages/_lib/folders.ts
@@ -2,11 +2,15 @@
 import { PrismaClient, Prisma } from '@prisma/client'
 import { v4 as uuidv4 } from "uuid";
 import { Folders } from '@/types/pagesFolder';
+import { allowedAdminAccess } from '@/lib/auth';
 
 
 const prisma = new PrismaClient();
 
 export const mxrFolderAdd = async (name: string, parent: string, order: number): Promise<Folders> => {
+    if (!allowedAdminAccess()) {
+        throw new Error('Access Denied');
+    }
     const folder = await prisma.folders.create({
         data: {
             id: uuidv4(),
@@ -19,6 +23,9 @@ export const mxrFolderAdd = async (name: string, parent: string, order: number):
 }
 
 export const mxrFolderGet = async (id: string): Promise<Folders | null> => {
+    if (!allowedAdminAccess()) {
+        throw new Error('Access Denied');
+    }
     const folder = await prisma.folders.findUnique({
         where: {
             id: id
@@ -28,6 +35,9 @@ export const mxrFolderGet = async (id: string): Promise<Folders | null> => {
 }
 
 export const mxrFolderGetAll = async (): Promise<Folders[]> => {
+    if (!allowedAdminAccess()) {
+        throw new Error('Access Denied');
+    }
     const folders = await prisma.folders.findMany({
         where: {
             OR: [
@@ -44,6 +54,9 @@ export const mxrFolderGetAll = async (): Promise<Folders[]> => {
 };
 
 export const mxrFolderGetChildren = async (id: string): Promise<Folders[]> => {
+    if (!allowedAdminAccess()) {
+        throw new Error('Access Denied');
+    }
     const folders = await prisma.folders.findMany({
         where: {
             parentId: id
@@ -55,7 +68,11 @@ export const mxrFolderGetChildren = async (id: string): Promise<Folders[]> => {
     return folders;
 }
 
+// TODO create definitions for deletedFolders and deletedPages
 export const mxrFolderDelete = async (id: string): Promise<{ deletedFolders: any, deletedPages: any }> => {
+    if (!allowedAdminAccess()) {
+        throw new Error('Access Denied');
+    }
     const deletedFolders = await prisma.folders.deleteMany({
         where: {
             OR: [
@@ -73,6 +90,9 @@ export const mxrFolderDelete = async (id: string): Promise<{ deletedFolders: any
 }
 
 export const mxrFolderDeleteSingle = async (id: string): Promise<Folders> => {
+    if (!allowedAdminAccess()) {
+        throw new Error('Access Denied');
+    }
     const folder = await prisma.folders.delete({
         where: {
             id: id
@@ -82,6 +102,9 @@ export const mxrFolderDeleteSingle = async (id: string): Promise<Folders> => {
 }
 
 export const mxrFolderUpdate = async (id: string, name: string, order: number): Promise<Folders> => {
+    if (!allowedAdminAccess()) {
+        throw new Error('Access Denied');
+    }
     console.log("mxrFolderUpdate", id, name, order)
     const folder = await prisma.folders.update({
         where: {
